Change the map from the Qt6 OSM plugin to MapLibre with OSM.
When Qt upgrades from 5 to 6, it droped the mapbox-gl plugin
from QtLocations.  After that, AGL uses the osm plugin.
It shows the "API Key Required" message on each map tile.
It could not erase.
Maplibre is an alternative solution for the mapbox-gl plugin
and the osm plugin for Qt6.

This patch enables the maplibre-native-qt plugin in mominavi.

Bug-AGL: SPEC-5540

Change-Id: Ie06f38f0baba7a683750f0e117ccd2c6cae61a19
Signed-off-by: Naoto Yamaguchi <naoto.yamaguchi@aisin.co.jp>